Compose: use absolute file paths in volumes.
This is needed because we need them to be files, which interferes with
docker's habit of interpreting relative paths as absolute.
The drawback is that paths are now relative to the PWD, not the location of
the compose file. This might not be a bad thing, but can lead to
unexpected side effects.